migrate sub-account state management to zustand
