+++ packages/frontend/app/routes/subaccounts/subaccounts.module.css
@@ -2,7 +2,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    justify-content: center;
+    padding-top: 5vh;
     height: 100%;
     width: 100%;
     font-family: var(--font-family-main);
